+++ examples/audit/log_custom_schema.go
@@ -9,6 +9,7 @@ import (
 
 	"examples/audit/util"
 
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/audit"
 )
@@ -19,11 +20,16 @@ func main() {
 		log.Fatal("Unauthorized: No token present")
 	}
 
+	config, err := pangea.NewConfig(
+		option.WithToken(token),
+		option.WithDomain(os.Getenv("PANGEA_DOMAIN")),
+	)
+	if err != nil {
+		log.Fatal("failed to create config")
+	}
+
 	auditcli, err := audit.New(
-		&pangea.Config{
-			Token:  token,
-			Domain: os.Getenv("PANGEA_DOMAIN"),
-		},
+		config,
 		audit.WithCustomSchema(util.CustomSchemaEvent{}),
 	)
 	if err != nil {