+++ examples/audit/search_custom_schema.go
@@ -10,6 +10,7 @@ import (
 	"os"
 	"time"
 
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/service/audit"
 )
@@ -20,13 +21,18 @@ func main() {
 		log.Fatal("Unauthorized: No token present")
 	}
 
+	config, err := pangea.NewConfig(
+		option.WithToken(token),
+		option.WithDomain(os.Getenv("PANGEA_DOMAIN")),
+		option.WithPollResultTimeout(60*time.Second),
+		option.WithQueuedRetryEnabled(true),
+	)
+	if err != nil {
+		log.Fatal("failed to create config")
+	}
+
 	auditcli, err := audit.New(
-		&pangea.Config{
-			Token:              token,
-			Domain:             os.Getenv("PANGEA_DOMAIN"),
-			PollResultTimeout:  60 * time.Second,
-			QueuedRetryEnabled: true,
-		},
+		config,
 		audit.WithCustomSchema(util.CustomSchemaEvent{}),
 	)
 	if err != nil {